Reduce duplicate in integration tests (#591)
⚠️ 🚧 

As part of the 1.0 release, we removed service_account support from
setup-gcloud, but many of the integration tests were actually duplicates
now, since they used to test service_account. This updates the
integration tests to run all the tests in a single job (reducing the
number of times we install NPM packages and download gcloud).